'refactoring database, truncating inputs'

+++ paraquick-models/ParaquickMessage.cs
@@ -4,12 +4,12 @@ namespace com.paralib.paraquick.Models
 {
 	public partial class ParaquickMessage
 	{
-		public int Id { get; set;}
-		public int SessionId { get; set;}
+		public long Id { get; set;}
+		public long SessionId { get; set;}
 		public int MessageSetSequence { get; set;}
 		public int MessageSequence { get; set;}
-		public int MessageTypeId { get; set;}
-		public int ApplicationEntityId { get; set;}
+		public long MessageTypeId { get; set;}
+		public long ApplicationEntityId { get; set;}
 		public string RequestId { get; set;}
 		public string RequestXml { get; set;}
 		public DateTime RequestDate { get; set;}
