+++ paraquick-models/ParaquickSession.cs
@@ -4,12 +4,12 @@ namespace com.paralib.paraquick.Models
 {
 	public partial class ParaquickSession
 	{
-		public int Id { get; set;}
-		public int CompanyId { get; set;}
+		public long Id { get; set;}
+		public long CompanyId { get; set;}
 		public string Ticket { get; set;}
 		public DateTime CreateDate { get; set;}
 		public DateTime? StartDate { get; set;}
 		public DateTime? EndDate { get; set;}
-		public int StatusId { get; set;}
+		public long StatusId { get; set;}
 	}
 }
