+++ paraquick-models/ParaquickSessionError.cs
@@ -4,8 +4,8 @@ namespace com.paralib.paraquick.Models
 {
 	public partial class ParaquickSessionError
 	{
-		public int Id { get; set;}
-		public int SessionId { get; set;}
+		public long Id { get; set;}
+		public long SessionId { get; set;}
 		public DateTime Date { get; set;}
 		public string Message { get; set;}
 	}
